+++ src/promptwatch/utils.py
@@ -1,3 +1,5 @@
+import datetime
+from decimal import Decimal
 import inspect
 import types
 def find_the_caller_in_the_stack(name:str=None, type:type = None):
@@ -11,8 +13,44 @@ def find_the_caller_in_the_stack(name:str=None, type:type = None):
             caller_frame = caller_frame.f_back
 
 
-def is_primitive_type(val):
-    return val.__class__.__module__=="builtins" and not type(val) is type and not type(val) is dict
+def is_primitive_type(type_val):
+    return type_val in [int, float, str, bool, Decimal, type(None)]
+
+
+def copy_dict_serializable_values(dict_value):
+    res={}
+    if isinstance(dict_value, dict):
+        for key, value in dict_value.items():
+            if  is_primitive_type(type(value)):
+                res[key] = value
+            elif isinstance(value, dict):
+                res[key] = copy_dict_serializable_values(value)
+            elif isinstance(value, list):
+                res[key] = copy_list_serializable_values(value)
+            elif isinstance(value, datetime.datetime):
+                res[key] = value.isoformat()
+        return res
+    else:
+        raise ValueError(f"Expected dict. Got: {dict_value}")
+
+def copy_list_serializable_values(list_value):
+    if isinstance(list_value, list):
+        value=[]
+        for i, item in enumerate(list_value):
+            if is_primitive_type(item):
+                value[i] = item
+            elif isinstance(item, list):
+                value[i] = copy_list_serializable_values(item)
+            elif isinstance(item, dict):
+                value[i] = copy_dict_serializable_values(item)
+            elif isinstance(value, datetime.datetime):
+                value[i] = value.isoformat()
+        return value
+    else:
+        raise ValueError(f"Expected list. Got: {value}")
+
+
+
 
 
 def wrap_a_method(object, function_name, decorator):